% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsServer.R
\name{get_plot_columns}
\alias{get_plot_columns}
\title{Get the columns to use in the plot based on the selected chart}
\usage{
get_plot_columns(selected_chart)
}
\arguments{
\item{selected_chart}{A string representing the selected chart type. Possible values are:
- "by_reporter_type"
- "by_reporter"
- "by_reporter_region"
- "by_report_seriousness"
- "by_age_group"
- "by_sex"}
}
\value{
A vector of column names to be used in the plot corresponding to the selected chart.
}
\description{
This function returns the specific columns to be used in the plot depending on the selected chart type.
}
