% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsServer.R
\name{stacked_bars}
\alias{stacked_bars}
\title{Add plotly stacked bar chart}
\usage{
stacked_bars(columns_to_use)
}
\arguments{
\item{columns_to_use}{A vector of column names to be used in the plot. This should be a subset of the columns 
available in the dataset, typically returned by the `get_plot_columns` function.}
}
\value{
A Plotly stacked bar chart.
}
\description{
This function creates a stacked bar chart using Plotly based on the selected columns.
The plot shows the distribution of the selected categories over the years.
}
